Refactored the existing UserService class into an interface and add implementation for it.
