+++ src/main/java/com/example/demo/services/UserService.java
@@ -1,31 +1,14 @@
 package com.example.demo.services;
 
 import com.example.demo.models.User;
-import com.example.demo.repositories.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.Cacheable;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@Service
-@RequiredArgsConstructor
-public class UserService {
-    private final UserRepository userRepository;
+public interface UserService {
 
-    @Cacheable("users")
-    public List<User> listUsers() {
-        return userRepository.findAll();
-    }
+    List<User> listUsers();
 
-    @CacheEvict(value = "users", allEntries = true)
-    public void saveUser(User user) {
-        userRepository.save(user);
-    }
+    void saveUser(User user);
 
-    @CacheEvict(value = "users", allEntries = true)
-    public void deleteUser(Long id) {
-        userRepository.deleteById(id);
-    }
+    void deleteUser(Long id);
 }
