+++ src/main/java/com/example/demo/services/impl/UserServiceImpl.java
@@ -1,0 +1,36 @@
+package com.example.demo.services.impl;
+
+import com.example.demo.models.User;
+import com.example.demo.repositories.UserRepository;
+import com.example.demo.services.UserService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class UserServiceImpl implements UserService {
+
+    private final UserRepository userRepository;
+
+    @Override
+    @Cacheable("users")
+    public List<User> listUsers() {
+        return userRepository.findAll();
+    }
+
+    @Override
+    @CacheEvict(value = "users", allEntries = true)
+    public void saveUser(User user) {
+        userRepository.save(user);
+    }
+
+    @Override
+    @CacheEvict(value = "users", allEntries = true)
+    public void deleteUser(Long id) {
+        userRepository.deleteById(id);
+    }
+}
